requirements.txt
Pillow==7.1.1
xlrd==1.2.0
